requirements.txt
asgiref==3.5.2
attrs==22.1.0
click==8.1.3
Django==4.1.2
djangorestframework==3.14.0
exceptiongroup==1.0.0
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.5
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
wait-for-it==2.2.1
